+++ src/main/java/com/nebulaparfums/nebula_parfums/service/ProveedorService.java
@@ -1,5 +1,6 @@
 package com.nebulaparfums.nebula_parfums.service;
 
+import com.nebulaparfums.nebula_parfums.exception.ResourceNotFoundException;
 import com.nebulaparfums.nebula_parfums.model.Proveedor;
 import com.nebulaparfums.nebula_parfums.repository.IProveedorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +22,7 @@ public class ProveedorService implements IProveedorService{
 
     @Override
     public Proveedor getProveedorById(Integer id) {
-        Proveedor proveedor = iProveedorRepository.findById(id).get();
+        Proveedor proveedor = iProveedorRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("No se encontro el proveedor"));
         return proveedor;
     }
 
@@ -32,7 +33,11 @@ public class ProveedorService implements IProveedorService{
 
     @Override
     public void deleteProveedor(Integer id) {
-        iProveedorRepository.deleteById(id);
+        if (iProveedorRepository.existsById(id)) {
+            iProveedorRepository.deleteById(id);
+        }else {
+            throw new ResourceNotFoundException("No se encontro el proveedor");
+        }
     }
 
     @Override
